Refactor: Image Component 병합
